🧹 seperate sqlite
